Made Exercise abstract class to make the 3 input textfields and do stuff
